Create new janus client on every submission.
This is now required because the function to get the hpke config
from the aggregators is now private and is automatically called
when creating a new client. But we cannot keep clients in our
roundconfig because then the python runtime doesn't like it.
Thus we have to recreate the client in each round.